requirements.txt
# Python dependencies (usually included with ComfyUI)
numpy>=1.21.0
Pillow>=9.0.0
torch>=1.12.0

# System dependency for MP4 export:
# Install ffmpeg: https://ffmpeg.org/download.html
# - Windows: choco install ffmpeg
# - macOS: brew install ffmpeg
# - Linux: sudo apt install ffmpeg
